requirements.txt
asttokens==3.0.0
branca==0.8.0
certifi==2024.8.30
charset-normalizer==2.0.12
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
fonttools==4.55.2
greenlet==3.1.1
idna==3.10
ipyleaflet==0.19.2
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.2
redis==5.2.1
requests==2.27.1
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
spotipy==2.23.0
SQLAlchemy==1.4.37
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
traitlets==5.14.3
traittypes==0.2.1
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
widgetsnbextension==4.0.13
xyzservices==2024.9.0
